fix non-linux build stubs to match linux signatures
The bpffs stubs referenced the deleted TargetDetails type instead of
Info, and cpuCount returned int instead of uint64.

+++ internal/pkg/instrumentation/kernel/cpu_other.go
@@ -5,4 +5,4 @@
 
 package kernel
 
-func cpuCount() (int, error) { return 0, nil }
+func cpuCount() (uint64, error) { return 0, nil }
